--- Auto-Git Commit ---

+++ core/database.php
@@ -101,7 +101,7 @@ class database
         return $this->_connection->prepare($sql);
     }
      //Procesar consulta, sql=consulta,return=devolver resultados o solo true o false si se ejecuto la operacion
-    private function consulta($sql, $return)
+    private function consulta($sql, $return,$delete_cache=true)
     {
         /*foreach ($this->llamadas as $key => $ll) {
             if($ll['consulta']==$sql){
@@ -116,7 +116,9 @@ class database
             if ($return) {
                 $rows = $query->fetchAll();
             }else{
-                cache::delete_cache();
+                if($delete_cache){
+                    cache::delete_cache();
+                }
             }
 
         } catch (\PDOException $e) {
@@ -189,7 +191,7 @@ class database
         return $row;
     }
 
-    public function insert($table, $idname, $insert)
+    public function insert($table, $idname, $insert,$delete_cache=true)
     { //consulta insert
         $valor_primario = "";
         $image          = array();
@@ -213,7 +215,7 @@ class database
             $sql .= ($value == "true" || $value == "false") ? $value : "'" . str_replace("'", "\\'", $value) . "'";
         }
         $sql .= ")";
-        $row = $this->consulta($sql, false);
+        $row = $this->consulta($sql, false,$delete_cache);
         if($row){
             $last_id=$this->get_last_insert_id();
             if (count($image) > 0) {
@@ -227,7 +229,7 @@ class database
         return $row;
     }
 
-    public function update($table, $idname, $set, $where)
+    public function update($table, $idname, $set, $where,$delete_cache=true)
     { //consulta update
         $set   = self::process_multiple($set);
         $image = array();
@@ -259,7 +261,7 @@ class database
         }
         $sql .= ") ";
         if (count($where) > 0) {
-            $row = $this->consulta($sql, false);
+            $row = $this->consulta($sql, false,$delete_cache);
             if($row){
                 if (count($image) > 0) {
                     $this->process_image($image, $table, $idname, $where[$idname]);
@@ -275,7 +277,7 @@ class database
         }
     }
 
-    public function delete($table, $idname, $where)
+    public function delete($table, $idname, $where,$delete_cache=true)
     { //consulta delete
         $sql = "DELETE FROM " . self::$_prefix . $table;
         $sql .= " WHERE (TRUE";
@@ -284,7 +286,7 @@ class database
         }
         $sql .= ")";
         if (count($where) > 0) {
-            $row = $this->consulta($sql, false);
+            $row = $this->consulta($sql, false,$delete_cache);
             image::delete($table, '', $where[$idname]);
             file::delete($table, '', $where[$idname]);
             return $row;
